+++ kikuchipy/io/plugins/tests/test_emsoft_ebsd.py
@@ -1,0 +1,87 @@
+# -*- coding: utf-8 -*-
+# Copyright 2019-2020 The kikuchipy developers
+#
+# This file is part of kikuchipy.
+#
+# kikuchipy is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# kikuchipy is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with kikuchipy. If not, see <http://www.gnu.org/licenses/>.
+
+import os
+
+import dask.array as da
+from h5py import File
+import numpy as np
+import pytest
+
+from kikuchipy.io._io import load
+from kikuchipy.io.plugins.emsoft_ebsd import _check_file_format
+from kikuchipy.signals.util._metadata import metadata_nodes
+
+DIR_PATH = os.path.dirname(__file__)
+EMSOFT_FILE = os.path.join(
+    DIR_PATH, "../../../data/emsoft_ebsd/simulated_ebsd.h5"
+)
+
+
+class TestEMsoftEBSDReader:
+    def test_file_reader(self):
+        """Test correct data shape, axes properties and metadata."""
+        s = load(EMSOFT_FILE)
+
+        assert s.data.shape == (10, 10, 10)
+
+        assert s.axes_manager["dx"].scale == 70
+        assert s.axes_manager["dx"].units == "um"
+        assert s.axes_manager["x"].units == "px"
+
+        sem_node, ebsd_node = metadata_nodes(["sem", "ebsd"])
+        phase_node = "Sample.Phases.1"
+        desired_dict = {
+            f"{sem_node}.beam_energy": 20,
+            f"{ebsd_node}.manufacturer": "EMsoft",
+            f"{ebsd_node}.xpc": 4.86,
+            f"{phase_node}.space_group": 140,
+            f"{phase_node}.atom_coordinates.1.atom": 13,
+        }
+        for key, value in desired_dict.items():
+            assert s.metadata.get_item(key) == value
+
+    @pytest.mark.parametrize("scan_size", [10, (1, 10), (5, 2)])
+    def test_scan_size(self, scan_size):
+        """Scan size parameter works as expected."""
+        s = load(EMSOFT_FILE, scan_size=scan_size)
+
+        sy, sx = (10, 10)
+        if isinstance(scan_size, int):
+            desired_shape = (scan_size, sy, sx)
+        else:
+            desired_shape = scan_size + (sy, sx)
+
+        assert s.data.shape == desired_shape
+
+    def test_read_lazy(self):
+        """Lazy parameter works as expected."""
+        s = load(EMSOFT_FILE, lazy=True)
+
+        assert isinstance(s.data, da.Array)
+
+    def test_check_file_format(self, save_path_hdf5):
+        """Wrong file format raises an error."""
+        with File(save_path_hdf5, mode="w") as f:
+            g1 = f.create_group("EMheader")
+            g2 = g1.create_group("EBSD")
+            g2.create_dataset(
+                "ProgramName", data=np.array([b"EMEBSDD.f90"], dtype="S11"),
+            )
+            with pytest.raises(IOError, match=".* is not in EMsoft's format "):
+                _ = _check_file_format(f)